refactor(db): move the media content parts out of prompts.py
`prompts.py` defines the parts a prompt message can hold, and the fork had added a
hundred and thirty lines to it: two part classes with their validators, the media
union, four helpers, and a validator body on `PromptMessage`. That is the fork's
feature interleaved with definitions upstream owns.

It now lives in `media_parts.py`. `prompts.py` keeps what it cannot delegate: the
import, five lines naming the new parts in its `ContentPart` union, and a
three-line validator that calls `reject_media_on_non_user_role`. Seventeen lines
instead of a hundred and thirty.

There is no runtime import cycle. `prompts` imports the classes because its union
names them; `media_parts` needs `ContentPart` and `PromptChatTemplate` only to
annotate two helpers, so those imports are under TYPE_CHECKING and the annotations
are strings — without the quotes the module raises NameError on import, which is how
the first attempt failed.
