+++ src/phoenix/server/daemons/media_sweeper.py
@@ -19,6 +19,8 @@ from phoenix.db.types.media import (
 from phoenix.db.types.prompts import (
     PromptChatTemplate,
     PromptTemplate,
+)
+from phoenix.db.types.media_parts import (
     is_media_content_part,
     media_source,
 )
